requirements.txt
numpy~=1.19.5
pandas~=1.1.5
sklearn~=0.0
scikit-learn~=0.24.2
jass-kit
flask
gunicorn